Resolved potential issue in the test case
