+++ src/test/java/io/openliberty/tools/intellij/it/GradleSingleModMPSIDProjectTest.java
@@ -92,16 +92,6 @@ public class GradleSingleModMPSIDProjectTest extends SingleModMPProjectTestCommo
         }
     }
 
-    @Override
-    public String getCustomWLPPath() {
-        return "";
-    }
-
-    @Override
-    public String getTargetDir() {
-        return "";
-    }
-
     GradleSingleModMPSIDProjectTest() {
         // set the new locations for the test, not the original locations
         setProjectsDirPath(PROJECTS_PATH_NEW);
@@ -117,5 +107,6 @@ public class GradleSingleModMPSIDProjectTest extends SingleModMPProjectTestCommo
         setStartParams("--hotTests");
         setStartParamsDebugPort("--libertyDebugPort=9876");
         setProjectTypeIsMultiple(false);
+        setBuildDirectory("build");
     }
 }
