+++ src/test/java/io/openliberty/tools/intellij/it/MavenMPMultipleProjectTest.java
@@ -126,6 +126,7 @@ public class MavenMPMultipleProjectTest extends SingleModMPProjectTestCommon {
         setStartParams("-DhotTests=true");
         setStartParamsDebugPort("-DdebugPort=9876");
         setProjectTypeIsMultiple(true);
+        setBuildDirectory("target");
     }
 
     /**
@@ -151,15 +152,5 @@ public class MavenMPMultipleProjectTest extends SingleModMPProjectTestCommon {
     public void validateTestReportsExist() {
         TestUtils.validateTestReportExists(pathToITReport34, pathToITReport35);
         TestUtils.validateTestReportExists(pathToUTReport34, pathToUTReport35);
-    }
-
-    @Override
-    public String getCustomWLPPath() {
-        return "";
-    }
-
-    @Override
-    public String getTargetDir() {
-        return "";
     }
 }
